[BE] FEAT: 백엔드 서버 내 토큰 생성 및 쿠키 set 확인

+++ backend/src/main/java/proj/pet/auth/domain/jwt/JwtProperties.java
@@ -19,7 +19,7 @@ public class JwtProperties {
 	private String tokenName;
 
 	@Value("${cloud.oauth2.jwt.token.expiry}")
-	private Integer expiry;
+	private long expiry;
 
 	public Key getSigningKey() {
 		byte[] secretKeyBytes = DatatypeConverter.parseBase64Binary(this.secret);
